removed Image from imports, moved functions to layer_utils
